requirements.txt
python-dotenv>=1.0.0
numpy>=1.23
sounddevice>=0.4.6
scipy>=1.10
soundfile>=0.12.1
pydub>=0.25.1
TTS>=0.22.0
playwright>=1.42.0
torchcodec
tqdm>=4.66.0
